Move all other modules to aeslibs folder
