+++ __init__.py
@@ -1,5 +1,4 @@
 from . import aestimo
-from . import aestimo_eh
 from . import config
 from . import database
 from . import intersubband_optical_transitions
