+++ aeslibs/ddgelectron_driftdiffusion.py
@@ -50,16 +50,9 @@ import numpy as np
 from math import*
 from scipy import sparse as sp
 
-if __package__:  # explicit relative imports for using aestimo as a package (in python3)
-    from . import func_lib
-    from .func_lib import DDGphin2n,DDGphip2p,Ucompmass,Ucomplap,Ucompconst,Ubernoulli
-    from .aestimo_poisson1d import equi_np_fi222
-    from . import config
-else:
-    import func_lib
-    from func_lib import DDGphin2n,DDGphip2p,Ucompmass,Ucomplap,Ucompconst,Ubernoulli
-    from aestimo_poisson1d import equi_np_fi222
-    import config
+from .func_lib import DDGphin2n,DDGphip2p,Ucompmass,Ucomplap,Ucompconst,Ubernoulli
+from .aestimo_poisson1d import equi_np_fi222
+import config
     
 def DDGelectron_driftdiffusion(psi,xaxis,ng,p,ni,TAUN0,TAUP0,mun,fi_e,fi_h,model,Vt,idata):
